edición de las funciones coordinador

+++ Cursos/Cursos.java
@@ -11,8 +11,8 @@ abstract public class Cursos {
      */
     protected String codigo; // Código del curso
     protected String nombre; // Nombre del curso
-    protected int creditos; // Cantidad de créditos del curso
-    protected int horasLectivas; // Cantidad de horas de clase
+    protected short creditos; // Cantidad de créditos del curso
+    protected short horasLectivas; // Cantidad de horas de clase
     protected String[] requisitos; // Cursos que se deben aprobar primero
     protected String[] corequisitos; // Cursos a llevar en el mismo semestre
 
@@ -26,7 +26,7 @@ abstract public class Cursos {
      * @param req (String[]) Lista de requisitos que se deben aprobar primero
      * @param cor (String[]) Lista de cursos de deben ser llevados al mismo tiempo
      */
-    public Cursos(String cod, String nom, int creditos, int horasLectivas, String[] req, String[] cor) {
+    public Cursos(String cod, String nom, short creditos, short horasLectivas, String[] req, String[] cor) {
         this.codigo = cod;
         this.nombre = nom;
         this.creditos = creditos;
@@ -76,7 +76,7 @@ abstract public class Cursos {
      * 
      * @return creditos (String)
      */
-    public int getCreditos() {
+    public short getCreditos() {
         return creditos;
     }
 
@@ -94,7 +94,7 @@ abstract public class Cursos {
      * 
      * @return horasLectivas (short)
      */
-    public int getHorasLectivas() {
+    public short getHorasLectivas() {
         return horasLectivas;
     }
