+++ Cursos/VirtualSincronico.java
@@ -31,7 +31,7 @@ public class VirtualSincronico extends Cursos {
      * @param pla  (String)
      */
     public VirtualSincronico(
-            String cod, String nom, int cre, int hl, String[] req,
+            String cod, String nom, short cre, short hl, String[] req,
             String[] cor, String[] dias, String hIn, String hFn, String pla) {
         super(cod, nom, cre, hl, req, cor);
         this.dias = dias;
